requirements.txt
amqp==5.0.5
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.2
importlib-metadata==3.7.3
ipdb==0.13.7
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kombu==5.0.2
nltk==3.5
numpy==1.20.1
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
